moved functions to main app
